fixed up threading! :zigfast:

+++ src/voxel/types.zig
@@ -3,7 +3,7 @@ const math = @import("math");
 
 usingnamespace math.glm;
 
-pub const VoxelTypeId = u8;
+pub const VoxelTypeId = u16;
 
 pub const ByteCoords = math.vector.Vector(u8, 3);
 pub const Coords = math.vector.Vector(i32, 3);
